Create br-profile-translation
still not sure if Switch is being sold in Portuguese countries. Brazil especially. but i wanted to add a translation anyway. also, put in something i forgot to put in a discord message (which is the translation for "unset"). also forgot about "Head", "Clothes", "Shoes", and "Splatoon 2 Profile" in the message so i'm going to put that here also so i don't forget or whatever.